autocreation of memberships

+++ app/models/membership.rb
@@ -2,38 +2,48 @@ class Membership < ActiveRecord::Base
   belongs_to :group
   belongs_to :user
 
-  validate :is_unique_in_group
-  before_save :default_values
-  after_create :create_ancestors
-  before_destroy :destroy_descendants
+  before_create :is_unique_in_group
+  after_create :create_parent_membership, :create_child_memberships
+  after_destroy :destroy_child_memberships
 
   def is_unique_in_group
-    if self.group.memberships.exists?(user_id: self.user_id)
+    if self.group.memberships.where(user_id: self.user_id).count > 1
       errors[:base].push("A user may have only one membership in a given group.")
     end
   end
 
   def default_values
-    if self.is_admin === nil
+    if !self.is_admin
       self.is_admin = "false"
+      self.save!
     end
   end
 
-  def create_ancestors
-    if self.group.parent
-      if !self.group.parent.memberships.exists?(user_id: self.user_id)
-        self.class.create!(group_id: self.group.parent_id, user_id: self.user_id, is_admin: "false")
+  def create_parent_membership
+    return if !self.group.parent
+    return if self.group.parent.memberships.exists?(user_id: self.user_id)
+    membership = self.group.parent.memberships.new
+    membership.user_id = self.user_id
+    membership.is_admin = false
+    membership.save! if membership.valid?
+  end
+
+  def create_child_memberships
+    return if !self.is_admin
+    self.group.children.each do |subgroup|
+      @m = subgroup.memberships.find_by(user_id: self.user_id)
+      if !@m
+        @m = subgroup.memberships.new
       end
+      @m.user_id = self.user_id
+      @m.is_admin = true
+      @m.save!
     end
   end
 
-  def destroy_descendants
-    if self.group.children
-      self.group.children.collect{|c| c.memberships}.flatten.each do |child|
-        if child.user_id == self.user_id
-          child.destroy!
-        end
-      end
+  def destroy_child_memberships
+    self.group.children.collect{|c| c.memberships}.flatten.each do |child|
+      child.destroy! if child.user_id == self.user_id
     end
   end
